branch ufscycle - changes needed to cycle ufs

+++ scripts/exrtofs_glo_analysis.sh
@@ -5,18 +5,12 @@
 # Script name:         exrtofs_glo_analysis.sh                                #
 # Script description:                                                         #
 #                                                                             #
-# Author:        Ilya Rivin      Org: NP23         Date: 2010-07-30           #
-#                                                                             #
 # Abstract: This script generates the analysis fields                         #
 #           for the RTOFS_GLO Ocean model                                     #
 #                                                                             #
 # Sub-scripts called:                                                         #
-#   rtofs_runstaging.sh - gets input files from COMIN FIX and PARM            #
-#   rtofs_submit.sh - submits the RTOFS HYCOM simulation                      #
-#   rtofs_tmp2com.sh - copies the products to COMOUT                          #
-#                                                                             #
-# Script history log:                                                         #
-# 2010-07-30  Ilya Rivin                                                      #
+#   rtofs_abort.sh - Error handling for model crashes                         #
+#   rtofs_combine_nc.sh - (via cfp) combine restarts and product files        #
 #                                                                             #
 ###############################################################################
 set -xa
@@ -29,16 +23,16 @@ msg="RTOFS_GLO_ANALYSIS JOB has begun on $(hostname) at $(date)"
 postmsg "$msg"
 
 # --------------------------------------------------------------------------- #
-# 0. set some defaults
-
-typeset -Z5 SSSSS
-export fcstdays=${fcstdays:-2}
-export inputgrid=${inputgrid:-navy_0.08}
-
-# --------------------------------------------------------------------------- #
-# 1. Get initial conditions (restart files)
+# 0. Get initial conditions (restart files)
 #     if this is a restart, then find latest restart files in DATArestart
 #     if this is not a restart, then find restart files in COMin
+
+# THIS IS RESTART LOGIC FOR RTOFS V2.5 (HYCOM MODEL)
+# IT NEEDS TO BE UPDATED FOR RTOFS V3.0 (UFS MODEL)
+
+wearerunningv2=0
+if [ $wearerunningv2 -eq 1 ]
+then
 
   test -f restart_in.a && rm -f restart_in.a
   test -f restart_in.b && rm -f restart_in.b
@@ -79,7 +73,7 @@ export inputgrid=${inputgrid:-navy_0.08}
       ln -sf ${rfile%.b}.a $DATA/restart_in.a
       ln -sf $DATArestart/cice.restart.${YYYY}-${MM}-${DD}-${SSSSS} $DATA/cice.restart_in
       restart_found=yes
-      echo INFO - restarting simulation from $latestdate using $rfile 
+      echo INFO - restarting simulation from $latestdate using $rfile
     fi
 
     if [ $restart_found = no ]
@@ -110,7 +104,7 @@ export inputgrid=${inputgrid:-navy_0.08}
 
     if [ -s ${COMIN}/${HYCOMrestTplate}.a ] && \
        [ -s ${COMIN}/${HYCOMrestTplate}.b ] && \
-       [ -s ${COMIN}/${CICErestTplate} ] 
+       [ -s ${COMIN}/${CICErestTplate} ]
     then
       ln -s -f ${COMIN}/${HYCOMrestTplate}.a restart_in.a
       ln -s -f ${COMIN}/${HYCOMrestTplate}.b restart_in.b
@@ -135,73 +129,215 @@ export inputgrid=${inputgrid:-navy_0.08}
     $USHrtofs/${RUN}_abort.sh "FATAL ERROR: $job Missing Restart File" \
         "No restart_in.[ab] or cice.restart_in in $DATA" 911
   fi
+fi # wearerunningv2=1
 
 # --------------------------------------------------------------------------- #
-# 2. get input files, cice forcing may need to be regenerated
+# 1. copy in files from parm to top level
 
-  ksh ${USHrtofs}/${RUN}_runstaging.sh
-  if [ $RESTART = YES ]
-  then # create new cice files to this start time
-    rm -f cice.??????.?
-    export iday=$($USHrtofs/rtofs_date_normal2hycom.sh $startdate)
-    $USHrtofs/${RUN}_iceforcing.sh
-  fi
+adate=${PDYm1:0:4}-${PDYm1:4:2}-${PDYm1:6:2}-00000
+for pfile in ice_in; do
+cp $PARMrtofs/$pfile ./ice_in
+  sed -i -e "s/&YMDS.nc/$adate.nc/" -e "s/&adjust_aice/none/" ./ice_in
+done
+
+for pfile in data_table datm_in datm.streams diag_table fd_ufs.yaml input.nml noahmptable.tbl ufs.configure
+do
+  cp $PARMrtofs/$pfile .
+done
+  sed -i -e "s/&startup_continue/continue/" ./ufs.configure
+
+# model_configure created on the fly
+for pfile in model_configure
+do
+  cp $PARMrtofs/$pfile .
+  sed -i -e  "s/&YYYY/${PDYm1:0:4}/g" -e "s/&MM/${PDYm1:4:2}/g" -e "s/&DD/${PDYm1:6:2}/g" -e "s/&HH/00/g" -e "s/&NH/24/g" ./model_configure
+done
 
 # --------------------------------------------------------------------------- #
-# 3. Run analysis
+# 2a. link in fix files to top-level
 
-  ${USHrtofs}/${RUN}_submit.sh
+#for ffile in cice_model.res.nc grid_cice_NEMS_mx008.nc kmtu_cice_NEMS_mx008.nc mesh.mx008.nc
+for ffile in grid_cice_NEMS_mx008.nc kmtu_cice_NEMS_mx008.nc mesh.mx008.nc
+do
+  ln -sf $FIXrtofs/$ffile .
+done
 
-  ok="unknown"
-  test -s ${DATA}/summary_out && ok=$(tail -1 ${DATA}/summary_out)
-  if [ "$ok" = "normal stop" ]
-  then
-    modelstatus=0
-  else
-    modelstatus=1
-  fi
-
-#
 # --------------------------------------------------------------------------- #
-# 4. If run ran to completion then copy restart and archive files to COMOUT
+# 2b. link or copy fix and parm files to INPUT
+# fix depth 0p08 
 
-# if model ran, then copy restart files that correspond to the times n00 and n-06 to comout
-#   n00 - for forecast step1
-#   n-06 - for PDYp1 incup step
-  if [ $modelstatus = 0 ]
-  then
-      if [ $SENDCOM = 'YES' ]
-      then
-        ${USHrtofs}/${RUN}_tmp2com.sh
-      fi
-      for rfile in $(ls ${DATArestart}/restart_*.b)
-      do
-        cdate=$(${USHrtofs}/rtofs_date4restart.sh $rfile)
-        YYYY=$(echo $cdate | cut -c1-4)
-        MM=$(echo $cdate | cut -c5-6)
-        DD=$(echo $cdate | cut -c7-8)
-        HH=$(echo $cdate | cut -c9-10)
-        SSSSS=$(expr $HH \* 3600)
-        LEAD=$($NHOUR ${cdate} ${PDY}${mycyc})
-        if [[ $LEAD -eq 00 || $LEAD -eq -06 ]]
-        then
-          HYCOMrestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart
-          CICErestTplate=${RUN}_${modID}.t${mycyc}z.n${LEAD}.restart_cice
-          cp -p $rfile ${COMOUT}/${HYCOMrestTplate}.b
-          cp -p ${rfile%.b}.a ${COMOUT}/${HYCOMrestTplate}.a
-          cp -p ${DATArestart}/cice.restart.${YYYY}-${MM}-${DD}-${SSSSS} ${COMOUT}/${CICErestTplate}
-        fi
-      done
-    echo "done" >$COMOUT/${RUN}_${modID}.t${mycyc}z.anal.log
+mkdir INPUT
+#for ffile in chl_mom6.nc depth_GLBb0.08_09m11ob2_mom6.nc grid_spec.nc mesh.datm.3072x1536.nc mom6_vgrid.nc ocean_hgrid.nc ocean_mosaic.nc runoff.daitren.clim.0.08deg.nc sss_mom6.nc tidal_amplitude.nc
+for ffile in chl_mom6.nc grid_spec.nc mesh.datm.3072x1536.nc mom6_vgrid.nc regional.mom6.nc ocean_mosaic.nc runoff.daitren.clim.0.08deg.nc sss_mom6.nc tidal_amplitude.nc
+do
+  ln -sf $FIXrtofs/$ffile INPUT/.
+done
+  ln -s $FIXrtofs/depth_GLB.0p08_09m11ob2_mom6.nc INPUT/depth_GLBb0.08_09m11ob2_mom6.nc
 
-  else
-    $USHrtofs/${RUN}_abort.sh "FATAL ERROR: $job Abnormal model exit" \
-       "problem with analysis model run - return code $modelstatus" $modelstatus
-  fi
+for pfile in MOM_input
+do
+  cp $PARMrtofs/${pfile} .
+  sed -i -e "s/ODA_INCUPD = &Value /ODA_INCUPD = False/" ./MOM_input
+  mv MOM_input INPUT/
+done
+for pfile in MOM_layout MOM_override
+do
+  cp $PARMrtofs/$pfile INPUT
+done
+
+# --------------------------------------------------------------------------- #
+# 3. Populate INPUT directory with pdym1 restart and forcing
+
+# check that MOM.res files exist
+for momres in $(ls $COMIN/RESTART/${PDYm1}.000000.MOM.res*nc*)
+do
+   fn=$(basename $momres | cut -d. -f3-)
+   ln -s $momres INPUT/$fn
+done
+
+# forcing (for this time period)  (change datm.streams when changing)
+ln -s $COMIN/../forcing/$PDY/gfs.2025121400-2025122218_positive.nc INPUT/gfs.forcing.files.nc
+
+# iced with right date
+icedate=$(echo $PDYm1 | cut -c1-4)-$(echo $PDYm1 | cut -c5-6)-$(echo $PDYm1 | cut -c7-8)-00000
+ln -s $COMIN/RESTART/iced.${icedate}.nc INPUT/iced.${icedate}.nc
+echo INPUT/iced.${icedate}.nc > ice.restart_file
+
+#datm files 
+ln -s $COMIN/datm.gfs.datm.r.${icedate}.nc .
+echo datm.gfs.datm.r.${icedate}.nc > rpointer.atm
+
+ln -s $COMIN/RESTART/datm.gfs.cpl.r.${icedate}.nc INPUT/.
+echo  INPUT/datm.gfs.cpl.r.${icedate}.nc > rpointer.cpl
+
+# --------------------------------------------------------------------------- #
+# 4. Create other directories
+
+mkdir history RESTART MOM6_OUTPUT
+
+# --------------------------------------------------------------------------- #
+# 5. get executable
+
+cp $EXECrtofs/ufs_model.x fv3.exe
+
+mpiexec -np $NMPI --cpu-bind core  ./fv3.exe >> $pgmout 2>errfile
+rc=$?
+
+echo fv3.exe completed with exit code $rc
+
+if [ $rc -ne 0 ]
+then
+   echo we are saying goodbye
+   $USHrtofs/${RUN}_abort.sh "FATAL ERROR: $job Abnormal model exit" \
+      "problem with analysis model run - return code $modelstatus" $rc
+   exit
+fi
+
+# combine files (one restart and two archives) and copy to COMOUT
+rm -f cmdfile.cpout
+mkdir -p $COMOUT/RESTART $COMOUT/history $COMOUT/MOM6_OUTPUT
+insertedsleepcommands=15
+
+#Restarts
+echo "${USHrtofs}/rtofs_combine_nc.sh False $DATA/RESTART ${PDYm1}.180000.MOM.res.nc ${COMOUT}/RESTART/${PDYm1}.180000.MOM.res.nc > cmb.restart.${PDYm1}.180000.res.out" >> cmdfile.cpout
+for i in $(seq $insertedsleepcommands};do echo "sleep 10" >> cmdfile.cpout;done
+echo "${USHrtofs}/rtofs_combine_nc.sh False $DATA/RESTART ${PDY}.000000.MOM.res.nc ${COMOUT}/RESTART/${PDY}.000000.MOM.res.nc > cmb.restart.${PDY}.000000.res.out" >> cmdfile.cpout
+for i in $(seq $insertedsleepcommands};do echo "sleep 10" >> cmdfile.cpout;done
+for res in 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16
+do
+   echo "${USHrtofs}/rtofs_combine_nc.sh False $DATA/RESTART ${PDYm1}.180000.MOM.res_${res}.nc ${COMOUT}/RESTART/${PDYm1}.180000.MOM.res_${res}.nc > cmb.restart.${PDYm1}.180000.res_${res}.out" >> cmdfile.cpout
+   for i in $(seq $insertedsleepcommands};do echo "sleep 10" >> cmdfile.cpout;done
+   echo "${USHrtofs}/rtofs_combine_nc.sh False $DATA/RESTART ${PDY}.000000.MOM.res_${res}.nc ${COMOUT}/RESTART/${PDY}.000000.MOM.res_${res}.nc > cmb.restart.${PDY}.000000.res_${res}.out" >> cmdfile.cpout
+   for i in $(seq $insertedsleepcommands};do echo "sleep 10" >> cmdfile.cpout;done
+done
+
+#Diagnostics (can do better than this)
+for dfile in $(ls ocn*0000)
+do
+   dfil=$(echo $dfile | cut -d. -f1-2)
+   echo "${USHrtofs}/rtofs_combine_nc.sh $DATA $dfil ${COMOUT}/$dfil > cmb.$dfil.out" >> cmdfile.cpout
+   for i in $(seq $insertedsleepcommands};do echo "sleep 10" >> cmdfile.cpout;done
+done
+
+# copy singular files
+for ifile in $(ls history/iceh*.nc*)
+do
+   echo "cp -p -f $ifile $COMOUT/history" >> cmdfile.cpout
+done
+idate=$(echo $PDYm1 | cut -c1-4)-$(echo $PDYm1 | cut -c5-6)-$(echo $PDYm1 | cut -c7-8)-64800
+fdate=$(echo $PDY | cut -c1-4)-$(echo $PDY | cut -c5-6)-$(echo $PDY | cut -c7-8)-00000
+echo "cp -p -f RESTART/iced.${idate}.nc $COMOUT/RESTART" >> cmdfile.cpout
+echo "cp -p -f RESTART/iced.${fdate}.nc $COMOUT/RESTART" >> cmdfile.cpout
+echo "cp -p -f RESTART/datm.gfs.cpl.r.${idate}.nc $COMOUT/RESTART" >> cmdfile.cpout
+echo "cp -p -f RESTART/datm.gfs.cpl.r.${fdate}.nc $COMOUT/RESTART" >> cmdfile.cpout
+echo "cp -p -f datm.gfs.datm.r.${idate}.nc $COMOUT">> cmdfile.cpout
+echo "cp -p -f datm.gfs.datm.r.${fdate}.nc $COMOUT">> cmdfile.cpout
+
+# diagnostics and log files
+for logfile in ice_diag.d mediator.log atm.log
+do
+  echo "cp -p -f $logfile $COMOUT/analysis.$logfile" >> cmdfile.cpout
+done
+for momoutputfile in $(ls MOM6_OUTPUT)
+do
+   echo "cp -p -f MOM6_OUTPUT/$momoutputfile $COMOUT/MOM6_OUTPUT/analysis.$momoutputfile" >> cmdfile.cpout
+done
+
+chmod +x cmdfile.cpout
+mpiexec -np $NPROCS --cpu-bind verbose,core cfp ./cmdfile.cpout > cpout.out
+err=$? ; export err ; err_chk
+date
+
+#################################################
+msg="THE RTOFS_GLO_ANALYSIS JOB HAS ENDED NORMALLY on $(hostname) at $(date)."
+postmsg "$msg"
+
+exit
+
+# copy archive and history files to COMOUT
+rm -f cmdfile.cpout
+
+# ocean archives
+for ofile in $(ls ocn*.nc*)
+do
+   echo "cp -p -f $ofile $COMOUT" >> cmdfile.cpout
+done
+# ice history
+mkdir -p $COMOUT/history
+for ifile in $(ls history/iceh*.nc*)
+do
+   echo "cp -p -f $ifile $COMOUT/history" >> cmdfile.cpout
+done
+# restart
+mkdir -p $COMOUT/RESTART
+for rfile in $(ls RESTART/${PDYm1}.180000.MOM.res*.nc* RESTART/${PDY}.000000.MOM.res*.nc*)
+do
+  echo "cp -p -f $rfile $COMOUT/RESTART" >> cmdfile.cpout
+done
+idate=$(echo $PDYm1 | cut -c1-4)-$(echo $PDYm1 | cut -c5-6)-$(echo $PDYm1 | cut -c7-8)-64800
+fdate=$(echo $PDY | cut -c1-4)-$(echo $PDY | cut -c5-6)-$(echo $PDY | cut -c7-8)-00000
+echo "cp -p -f RESTART/iced.${idate}.nc RESTART/iced.${fdate}.nc $COMOUT/RESTART" >> cmdfile.cpout
+echo "cp -p -f RESTART/datm.gfs.cpl.r.${idate}.nc RESTART/datm.gfs.cpl.r.${fdate}.nc $COMOUT/RESTART" >> cmdfile.cpout 
+echo "cp -p -f datm.gfs.datm.r.${idate}.nc datm.gfs.datm.r.${fdate}.nc $COMOUT">> cmdfile.cpout
+
+# diagnostics and log files
+mkdir -p $COMOUT/MOM6_OUTPUT
+for logfile in ice_diag.d mediator.log atm.log
+do
+  echo "cp -p -f $logfile $COMOUT/analysis.$logfile" >> cmdfile.cpout
+done
+for momoutputfile in $(ls MOM6_OUTPUT)
+do
+   echo "cp -p -f MOM6_OUTPUT/$momoutputfile $COMOUT/MOM6_OUTPUT/analysis.$momoutputfile" >> cmdfile.cpout
+done
+
+chmod +x cmdfile.cpout
+mpiexec -np $NPROCS --cpu-bind verbose,core cfp ./cmdfile.cpout
+err=$? ; export err ; err_chk
+date
 
 #################################################
 msg="THE RTOFS_GLO_ANALYSIS JOB HAS ENDED NORMALLY on $(hostname) at $(date)."
 postmsg "$msg"
 
 ################## END OF SCRIPT #######################
-
